+++ src/main/java/nl/_42/restsecure/autoconfigure/RestAuthenticationFilter.java
@@ -1,20 +1,8 @@
 package nl._42.restsecure.autoconfigure;
 
-import static org.springframework.http.HttpMethod.POST;
-import static org.springframework.http.HttpStatus.UNAUTHORIZED;
-
-import java.io.IOException;
-
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import nl._42.restsecure.autoconfigure.authentication.CrowdUser;
-import nl._42.restsecure.autoconfigure.authentication.UserDetailsAdapter;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import nl._42.restsecure.autoconfigure.errorhandling.GenericErrorHandler;
-
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,12 +11,19 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.RememberMeServices;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import com.atlassian.crowd.integration.springsecurity.user.CrowdUserDetails;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+import static org.springframework.http.HttpMethod.POST;
+import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
 /**
  * Handles the login POST request. Tries to Authenticate the given user credentials using the auto configured {@link AuthenticationManager}.
@@ -46,57 +41,65 @@ public class RestAuthenticationFilter extends OncePerRequestFilter {
     private final GenericErrorHandler errorHandler;
     private final AntPathRequestMatcher requestMatcher;
     private final AuthenticationManager authenticationManager;
+    private final RememberMeServices rememberMeServices;
     private final ObjectMapper objectMapper;
 
-    public RestAuthenticationFilter(GenericErrorHandler handler, AuthenticationManager manager) {
+    public RestAuthenticationFilter(GenericErrorHandler handler, AuthenticationManager authenticationManager, RememberMeServices rememberMeServices) {
         this.errorHandler = handler;
         this.requestMatcher = new AntPathRequestMatcher("/authentication", POST.name());
-        this.authenticationManager = manager;
+        this.authenticationManager = authenticationManager;
+        this.rememberMeServices = rememberMeServices;
         this.objectMapper = new ObjectMapper();
     }
 
     @Override
     public void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws IOException, ServletException {
         if (requestMatcher.matches(request)) {
-            String loginFormJson = IOUtils.toString(request.getReader());
-            LoginForm form = objectMapper.readValue(loginFormJson, LoginForm.class);
-            UsernamePasswordAuthenticationToken token = new UsernamePasswordAuthenticationToken(form.username, form.password);
-            try {
-                log.info("Authenticating user: {}", form.username);
-                Authentication authentication = authenticationManager.authenticate(token);
-                authentication = convertIfNecessary(authentication);
-                SecurityContextHolder.getContext().setAuthentication(authentication);
-                request.setAttribute(LOGIN_FORM_JSON, loginFormJson);
-                chain.doFilter(request, response);
-            } catch (AuthenticationException ex) {
-                handleLoginFailure(response, ex);
-            }
+            doLogin(request, response, chain);
         } else {
             chain.doFilter(request, response);
         }
     }
 
-    private Authentication convertIfNecessary(Authentication authentication) {
-        if (!(authentication.getPrincipal() instanceof UserDetailsAdapter<?>)) {
-            log.info("Converting the Authentication principal to UserDetailsAdapter to enable @CurrentUser annotation.");
-            CrowdUserDetails crowdUserDetails = (CrowdUserDetails) authentication.getPrincipal();
-            return new UsernamePasswordAuthenticationToken(
-                    new UserDetailsAdapter<CrowdUser>(new CrowdUser(crowdUserDetails)),
-                    authentication.getCredentials(),
-                    authentication.getAuthorities());
+    private void doLogin(HttpServletRequest request, HttpServletResponse response, FilterChain chain) throws IOException, ServletException {
+        String loginFormJson = IOUtils.toString(request.getReader());
+        LoginForm form = objectMapper.readValue(loginFormJson, LoginForm.class);
+        UsernamePasswordAuthenticationToken token = new UsernamePasswordAuthenticationToken(form.username, form.password);
+
+        try {
+            log.info("Authenticating user: {}", form.username);
+
+            Authentication authentication = authenticationManager.authenticate(token);
+            SecurityContextHolder.getContext().setAuthentication(authentication);
+            request.setAttribute(LOGIN_FORM_JSON, loginFormJson);
+
+            if (rememberMeServices != null) {
+                rememberMeServices.loginSuccess(request, response, authentication);
+            }
+
+            chain.doFilter(request, response);
+        } catch (AuthenticationException ex) {
+            handleLoginFailure(request, response, ex);
         }
-        return authentication;
     }
 
-    private void handleLoginFailure(HttpServletResponse response, AuthenticationException exception) throws IOException {
+    private void handleLoginFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
+        log.warn("Authentication failure: {}", exception.getMessage());
+
         SecurityContextHolder.getContext().setAuthentication(null);
         errorHandler.respond(response, UNAUTHORIZED, SERVER_LOGIN_FAILED_ERROR);
-        log.warn("Authentication failure: {}", exception.getMessage());
+
+        if (rememberMeServices != null) {
+            rememberMeServices.loginFail(request, response);
+        }
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class LoginForm {
+
         public String username;
         public String password;
+
     }
+
 }