+++ src/main/java/nl/_42/restsecure/autoconfigure/WebSecurityAutoConfig.java
@@ -1,26 +1,21 @@
 package nl._42.restsecure.autoconfigure;
 
-import static org.springframework.http.HttpMethod.DELETE;
-import static org.springframework.security.core.context.SecurityContextHolder.MODE_INHERITABLETHREADLOCAL;
-import static org.springframework.security.web.csrf.CookieCsrfTokenRepository.withHttpOnlyFalse;
-import static org.springframework.util.Assert.notEmpty;
-
-import java.util.List;
-
-import javax.servlet.Filter;
-
 import nl._42.restsecure.autoconfigure.authentication.AbstractUserDetailsService;
 import nl._42.restsecure.autoconfigure.authentication.AuthenticationController;
+import nl._42.restsecure.autoconfigure.authentication.AuthenticationResultProvider;
+import nl._42.restsecure.autoconfigure.authentication.CurrentUserArgumentResolver;
+import nl._42.restsecure.autoconfigure.authentication.DefaultAuthenticationResultProvider;
+import nl._42.restsecure.autoconfigure.authentication.DefaultUserProvider;
+import nl._42.restsecure.autoconfigure.authentication.UserProvider;
+import nl._42.restsecure.autoconfigure.authentication.UserResolver;
 import nl._42.restsecure.autoconfigure.errorhandling.GenericErrorHandler;
 import nl._42.restsecure.autoconfigure.errorhandling.RestAccessDeniedHandler;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-
 import org.springframework.boot.autoconfigure.web.servlet.WebMvcAutoConfiguration;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
@@ -39,10 +34,21 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.authentication.AnonymousAuthenticationFilter;
+import org.springframework.security.web.authentication.RememberMeServices;
 import org.springframework.security.web.authentication.logout.HttpStatusReturningLogoutSuccessHandler;
 import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
 import org.springframework.security.web.csrf.CsrfTokenRepository;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import javax.servlet.Filter;
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.springframework.http.HttpMethod.DELETE;
+import static org.springframework.security.core.context.SecurityContextHolder.MODE_INHERITABLETHREADLOCAL;
+import static org.springframework.security.web.csrf.CookieCsrfTokenRepository.withHttpOnlyFalse;
 
 /**
  * Auto-configures Spring Web Security with a customized UserDetailsService for internal users storage or with crowd-integration-springsecurity for external crowd authentication.
@@ -54,7 +60,7 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
  */
 @Configuration
 @AutoConfigureAfter(WebMvcAutoConfiguration.class)
-@ComponentScan(basePackageClasses = {AuthenticationController.class, GenericErrorHandler.class})
+@ComponentScan(basePackageClasses = { AuthenticationController.class, GenericErrorHandler.class })
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(prePostEnabled = true)
 public class WebSecurityAutoConfig extends WebSecurityConfigurerAdapter {
@@ -67,20 +73,27 @@ public class WebSecurityAutoConfig extends WebSecurityConfigurerAdapter {
 
     @Autowired
     private RestAccessDeniedHandler accessDeniedHandler;
+
     @Autowired
     private GenericErrorHandler errorHandler;
+
     @Autowired(required = false)
-    private AbstractUserDetailsService<?> userDetailsService;
+    private AbstractUserDetailsService userDetailsService;
+
+    @Autowired(required = false)
+    private List<AuthenticationProvider> authProviders = new ArrayList<>();
+
     @Autowired(required = false)
     private RequestAuthorizationCustomizer authCustomizer;
+
     @Autowired(required = false)
     private HttpSecurityCustomizer httpCustomizer;
-    @Autowired(required = false)
-    private CustomAuthenticationProviders customAuthenticationProviders;
+
     @Autowired(required = false)
     private WebSecurityCustomizer webSecurityCustomizer;
+
     @Autowired(required = false)
-    private AuthenticationProvider crowdAuthenticationProvider;
+    private RememberMeServices rememberMeServices;
 
     /**
      * Adds a {@link BCryptPasswordEncoder} to the {@link ApplicationContext} if no {@link PasswordEncoder} bean is found already.
@@ -93,7 +106,27 @@ public class WebSecurityAutoConfig extends WebSecurityConfigurerAdapter {
         log.info("Adding default BCryptPasswordEncoder to ApplicationContext.");
         return new BCryptPasswordEncoder();
     }
-    
+
+    /**
+     * Adds an {@link UserProvider} to the {@link ApplicationContext} if no {@link UserProvider} bean is found already.
+     * @return UserMapper
+     */
+    @Bean
+    @ConditionalOnMissingBean(UserProvider.class)
+    public UserProvider userProvider() {
+        return new DefaultUserProvider();
+    }
+
+    /**
+     * Adds an {@link AuthenticationResultProvider} to the {@link ApplicationContext} if no {@link AuthenticationResultProvider} bean is found already.
+     * @return AuthenticationResultProvider
+     */
+    @Bean
+    @ConditionalOnMissingBean(AuthenticationResultProvider.class)
+    public AuthenticationResultProvider authenticationResultProvider() {
+        return new DefaultAuthenticationResultProvider();
+    }
+
     /**
      * {@inheritDoc}
      */
@@ -109,21 +142,18 @@ public class WebSecurityAutoConfig extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
         if (userDetailsService != null) {
-            log.info("Found userDetailService in ApplicationContext; configuring for local authentication store.");
+            log.info("Found UserDetailService in ApplicationContext.");
             auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder());
-        } else if (crowdAuthenticationProvider != null) {
-            log.info("Found crowd authenticationProvider in ApplicationContext; configuring for crowd authentication store.");
-            auth.authenticationProvider(crowdAuthenticationProvider);
+        } else if (!authProviders.isEmpty()) {
+            log.info("Found AuthenticationProvider(s) in ApplicationContext.");
+            authProviders.forEach((authProvider) -> {
+                log.info("\t Registering '{}' as authentication provider.", authProvider.getClass().getSimpleName());
+                auth.authenticationProvider(authProvider);
+            });
         } else {
             throw new IllegalStateException(
-                    "Cannot configure security; either an AbstractUserDetailsService bean must be provided "
-                            + "or crowd-integration-springsecurity.jar must be on the classpath.");
-        }
-        if (customAuthenticationProviders != null) {
-            List<AuthenticationProvider> providers = customAuthenticationProviders.get();
-            notEmpty(providers, "CustomAuthenticationProviders bean must return at least one AuthenticationProvider.");
-            log.info("Found customAuthenticationProviders bean in ApplicationContext; adding {} custom providers to web security.", providers.size());
-            providers.forEach(auth::authenticationProvider);
+                "Cannot configure security; either a UserDetailsService or AuthenticationProvider bean must be present."
+            );
         }
     }
 
@@ -165,7 +195,7 @@ public class WebSecurityAutoConfig extends WebSecurityConfigurerAdapter {
     }
    
     private Filter authenticationFilter() throws Exception {
-        return new RestAuthenticationFilter(errorHandler, authenticationManagerBean());
+        return new RestAuthenticationFilter(errorHandler, authenticationManagerBean(), rememberMeServices);
     }
     
     private ExpressionUrlAuthorizationConfigurer<HttpSecurity>.ExpressionInterceptUrlRegistry customize(
@@ -194,4 +224,18 @@ public class WebSecurityAutoConfig extends WebSecurityConfigurerAdapter {
         repository.setCookiePath("/");
         return repository;
     }
+
+    @Configuration
+    public static class WebSecurityMvcAutoConfig implements WebMvcConfigurer {
+
+        @Autowired
+        private UserResolver userResolver;
+
+        @Override
+        public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+            resolvers.add(new CurrentUserArgumentResolver(userResolver));
+        }
+
+    }
+
 }