+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/AuthenticationController.java
@@ -1,20 +1,11 @@
 package nl._42.restsecure.autoconfigure.authentication;
 
-import static java.util.Collections.singletonMap;
-
-import java.util.Map;
-import java.util.Set;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
-import org.springframework.security.web.csrf.CsrfToken;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * This controller implements the default /authentication and /authentication/handshake endpoints.
@@ -25,40 +16,20 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @RequestMapping("/authentication")
 public class AuthenticationController {
 
-    @Autowired(required = false)
-    private AuthenticationResultProvider authenticationResultProvider;
-    
+    private final AuthenticationResultProvider provider;
+
+    public AuthenticationController(AuthenticationResultProvider provider) {
+        this.provider = provider;
+    }
+
     @PostMapping
-    AuthenticationResult authenticate(@CurrentUser RegisteredUser user) {
-        return transform(user);
+    public AuthenticationResult authenticate(@CurrentUser RegisteredUser user) {
+        return current(user);
     }
 
     @GetMapping("/current")
-    AuthenticationResult current(@CurrentUser RegisteredUser user) {
-        return transform(user);
+    public AuthenticationResult current(@CurrentUser RegisteredUser user) {
+        return provider.toResult(user);
     }
 
-    private AuthenticationResult transform(RegisteredUser user) {
-        AuthenticationResult authentication;
-        if (authenticationResultProvider != null) {
-            authentication = authenticationResultProvider.toAuthenticationResult(user);
-        } else if (user instanceof CrowdUser) {
-            authentication = new CrowdUserResult((CrowdUser) user);
-        } else {
-            authentication = new AuthenticationResult() {
-                @Override
-                @JsonProperty
-                public String getUsername() {
-                    return user.getUsername();
-                }
-
-                @Override
-                @JsonProperty
-                public Set<String> getAuthorities() {
-                    return user.getAuthorities();
-                }
-            };
-        }
-        return authentication;
-    }
 }