+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/AuthenticationResult.java
@@ -12,7 +12,8 @@ public interface AuthenticationResult {
 
     @JsonProperty
     String getUsername();
-        
+
     @JsonProperty
     Set<String> getAuthorities();
+
 }