+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/RegisteredUser.java
@@ -12,7 +12,9 @@ public interface RegisteredUser {
 
     String getUsername();
     
-    String getPassword();
+    default String getPassword() {
+        return "";
+    }
     
     Set<String> getAuthorities();
     
@@ -31,4 +33,5 @@ public interface RegisteredUser {
     default boolean isEnabled() {
         return true;
     }
+
 }