+++ src/main/java/nl/_42/restsecure/autoconfigure/authentication/UserProvider.java
@@ -1,0 +1,9 @@
+package nl._42.restsecure.autoconfigure.authentication;
+
+import org.springframework.security.core.Authentication;
+
+public interface UserProvider {
+
+  RegisteredUser toUser(Authentication authentication);
+
+}