+++ src/test/java/nl/_42/restsecure/autoconfigure/authentication/AuthenticationControllerTest.java
@@ -1,9 +1,11 @@
 package nl._42.restsecure.autoconfigure.authentication;
 
 import static nl._42.restsecure.autoconfigure.RestAuthenticationFilter.SERVER_LOGIN_FAILED_ERROR;
+import static nl._42.restsecure.autoconfigure.test.RememberMeServicesConfig.REMEMBER_ME_HEADER;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.anonymous;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -15,10 +17,9 @@ import nl._42.restsecure.autoconfigure.test.AuthenticationResultProviderConfig;
 import nl._42.restsecure.autoconfigure.test.CredentialsExpiredUserConfig;
 import nl._42.restsecure.autoconfigure.test.CustomWebSecurityAndHttpSecurityConfig;
 import nl._42.restsecure.autoconfigure.test.FailingTwoFactorAuthenticationFilterConfig;
-import nl._42.restsecure.autoconfigure.test.InMemoryCrowdConfig;
-import nl._42.restsecure.autoconfigure.test.MockedCrowdAuthenticationProviderConfig;
 import nl._42.restsecure.autoconfigure.test.NoopPasswordEncoderConfig;
 
+import nl._42.restsecure.autoconfigure.test.RememberMeServicesConfig;
 import org.junit.Test;
 
 public class AuthenticationControllerTest extends AbstractApplicationContextTest {
@@ -69,6 +70,27 @@ public class AuthenticationControllerTest extends AbstractApplicationContextTest
     }
 
     @Test
+    public void authenticate_shouldSucceed_withRememberMeServices() throws Exception {
+        getWebClient(ActiveUserConfig.class, NoopPasswordEncoderConfig.class, RememberMeServicesConfig.class)
+                .perform(post("/authentication")
+                        .content("{\"username\": \"custom\", \"password\": \"password\"}"))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("authorities[0]").value("ROLE_ADMIN"))
+                .andExpect(jsonPath("username").value("username"))
+                .andExpect(header().string(REMEMBER_ME_HEADER, "success"));
+    }
+
+    @Test
+    public void authenticate_shouldFail_withRememberMeServices() throws Exception {
+        getWebClient(ActiveUserConfig.class, NoopPasswordEncoderConfig.class, RememberMeServicesConfig.class)
+                .perform(post("/authentication")
+                        .content("{\"username\": \"custom\", \"password\": \"other\"}"))
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("errorCode").value(SERVER_LOGIN_FAILED_ERROR))
+                .andExpect(header().string(REMEMBER_ME_HEADER, "failed"));
+    }
+
+    @Test
     public void authenticate_shouldFail_withCorrectCredentials_butBrcyptPasswordEncoder() throws Exception {
         getWebClient(ActiveUserConfig.class)
             .perform(post("/authentication")
@@ -103,35 +125,6 @@ public class AuthenticationControllerTest extends AbstractApplicationContextTest
             .andExpect(status().isUnauthorized())
             .andExpect(jsonPath("errorCode").value(SERVER_LOGIN_FAILED_ERROR));
     }
-    
-    @Test
-    public void authenticate_shouldSucceed_whenUsingInMemoryCrowdUsers() throws Exception {
-        getWebClient(InMemoryCrowdConfig.class, NoopPasswordEncoderConfig.class)
-            .perform(post("/authentication")
-                .content("{\"username\": \"janUser\", \"password\": \"secret\"}"))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("authorities[0]").value("ROLE_USER"))
-            .andExpect(jsonPath("username").value("janUser"));
-    }
-
-    @Test
-    public void authenticate_shouldFail_withIncorrectCredentials() throws Exception {
-        getWebClient(InMemoryCrowdConfig.class, NoopPasswordEncoderConfig.class)
-            .perform(post("/authentication")
-                .content("{\"username\": \"unknownUser\", \"password\": \"secret\"}"))
-            .andExpect(status().isUnauthorized())
-            .andExpect(jsonPath("errorCode").value(SERVER_LOGIN_FAILED_ERROR));
-    }
-    
-    @Test
-    public void authenticate_shouldSucceed_whenUsingMockedCrowdAuthenticationProvider() throws Exception {
-        getWebClient(InMemoryCrowdConfig.class, MockedCrowdAuthenticationProviderConfig.class)
-            .perform(post("/authentication")
-                .content("{\"username\": \"crowdUser\", \"password\": \"crowdPassword\"}"))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("authorities[0]").value("ROLE_USER"))
-            .andExpect(jsonPath("username").value("crowdUser"));
-    }
 
     @Test
     public void authenticate_shouldFail_withAuthenticationExceptionInSubsequentFilter() throws Exception {
@@ -141,4 +134,5 @@ public class AuthenticationControllerTest extends AbstractApplicationContextTest
                 .andExpect(status().isUnauthorized())
                 .andExpect(jsonPath("errorCode").value(SERVER_LOGIN_FAILED_ERROR));
     }
+
 }