+++ src/test/java/nl/_42/restsecure/autoconfigure/test/NullAuthenticationProviderConfig.java
@@ -6,11 +6,11 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.security.authentication.AuthenticationProvider;
 
 @Configuration
-public class NullCrowdAuthenticationProviderConfig {
+public class NullAuthenticationProviderConfig {
 
     @Bean
-    @Primary
-    public AuthenticationProvider crowdAuthenticationProvider() throws Exception {
+    public AuthenticationProvider crowdAuthenticationProvider() {
         return null;
     }
+
 }