+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+- REST authentication filter now supports `RememberMeServices`
+- Removed CROWD, separating library from any Authentication Provider implementation
+ * For CROWD we suggest using the `spring-boot-starter-crowd`
+
 ## [6.0.0] - 2019-05-07
 - Now using REST to connect crowd server.